+++ src/helpers/TransformHelpers.php
@@ -18,6 +18,46 @@ use spacecatninja\imagerx\services\ImagerService;
 class TransformHelpers
 {
     /**
+     * Parses quick transform syntax
+     */
+    public static function parseQuickTransforms(array $transform): array
+    {
+        if (count($transform) < 2) {
+            return [];
+        }
+
+        $width1 = (int)$transform[0];
+        $width2 = (int)$transform[1];
+        $defaults = $transform[2] ?? [];
+        $format = $transform[3] ?? null;
+
+        if (is_numeric($defaults)) {
+            $defaults = ['ratio' => $defaults];
+        }
+
+        if (is_string($defaults)) {
+            $defaults = ['format' => $defaults];
+        }
+
+        if (is_string($format)) {
+            $defaults['format'] = $format;
+        }
+
+        return [
+            array_merge(['width' => min($width1, $width2)], $defaults),
+            array_merge(['width' => max($width1, $width2)], $defaults),
+        ];
+    }
+
+    /**
+     * Checks if transforms is a quick syntax
+     */
+    public static function isQuickSyntax(?array $transforms): bool
+    {
+        return is_array($transforms) && count($transforms) >= 2 && isset($transforms[0], $transforms[1]) && is_int($transforms[0]) && is_int($transforms[1]);
+    }
+
+    /**
      * Resolves any callables in params
      *
      * TODO : Make recursive, it now only resolves callables at the top level
@@ -37,24 +77,45 @@ class TransformHelpers
                     $transform[$key] = $resolvedVal;
                 }
             }
-            
+
             $r[] = $transform;
         }
 
         return $r;
     }
-    
+
     /**
      * Fills in the missing transform objects
-     *
-     *
      */
     public static function fillTransforms(array $transforms): array
     {
+        if (count($transforms) < 2) {
+            return $transforms;
+        }
+
         $r = [];
 
         $attributeName = ImagerService::$transformConfig->fillAttribute;
         $interval = (int)ImagerService::$transformConfig->fillInterval;
+
+        if (ImagerService::$transformConfig->fillTransforms === 'auto') {
+            $autoFillCount = ImagerService::$transformConfig->autoFillCount;
+            $firstTransform = $transforms[0];
+            $lastTransform = $transforms[count($transforms) - 1];
+
+            if (isset($firstTransform[$attributeName], $lastTransform[$attributeName])) {
+                $lastSize = (int)$lastTransform[$attributeName];
+                $firstSize = (int)$firstTransform[$attributeName];
+                $diff = $lastSize - $firstSize;
+
+                if ($autoFillCount === 'auto') {
+                    $targetInterval = $lastSize > 1000 ? 200 : 100;
+                    $autoFillCount = min(6, max(($diff / $targetInterval) - 1, 0));
+                }
+
+                $interval = ceil($diff / ($autoFillCount + 1));
+            }
+        }
 
         $r[] = $transforms[0];
 
@@ -83,7 +144,7 @@ class TransformHelpers
 
         return $r;
     }
-    
+
     /**
      * Merges default transform object into an array of transforms
      *
@@ -99,11 +160,9 @@ class TransformHelpers
 
         return $r;
     }
-    
+
     /**
      * Normalizes format of transforms
-     *
-     *
      */
     public static function normalizeTransforms(array $transforms, Asset|ImagerAdapterInterface|string $image): array
     {
@@ -118,15 +177,13 @@ class TransformHelpers
 
     /**
      * Normalize transform object and values
-     *
-     *
      */
     public static function normalizeTransform(array $transform, Asset|ImagerAdapterInterface|string $image): array
     {
         if (isset($transform['mode'])) {
             $transform['mode'] = mb_strtolower($transform['mode']);
         }
-        
+
         // if resize mode is not crop or croponly, remove position
         if (isset($transform['mode'], $transform['position']) && (($transform['mode'] !== 'crop') && ($transform['mode'] !== 'croponly'))) {
             unset($transform['position']);
@@ -161,7 +218,7 @@ class TransformHelpers
         // if transform is in Craft's named version, convert to percentage
         if (isset($transform['position'])) {
             if (\is_array($transform['position']) && isset($transform['position']['x'], $transform['position']['y'])) {
-                $transform['position'] = ($transform['position']['x'] * 100) . ' ' . ($transform['position']['y'] * 100);
+                $transform['position'] = ($transform['position']['x'] * 100).' '.($transform['position']['y'] * 100);
             }
 
             if (isset(ImagerService::$craftPositionTranslate[(string)$transform['position']])) {
@@ -170,7 +227,7 @@ class TransformHelpers
 
             $transform['position'] = str_replace('%', '', (string)$transform['position']);
         }
-        
+
         // normalize padding
         if (isset($transform['pad'])) {
             $transform['pad'] = self::normalizePadding($transform['pad']);
@@ -197,12 +254,12 @@ class TransformHelpers
         if (is_int($val)) {
             return [$val, $val, $val, $val];
         }
-        
+
         if (is_string($val)) {
             $val = str_replace('px', '', $val);
             $val = explode(' ', $val);
         }
-        
+
         if (is_array($val)) {
             if (count($val) === 1) {
                 return [(int)$val[0], (int)$val[0], (int)$val[0], (int)$val[0]];
@@ -224,7 +281,7 @@ class TransformHelpers
                 return [(int)$val[0], (int)$val[1], (int)$val[2], (int)$val[3]];
             }
         }
-        
+
         return null;
     }
 }